split example
