+++ examples/simple_dynamo.py
@@ -1,7 +1,4 @@
-import dis
-import types
 
-from functools import partial
 from typing import List
 
 import paddle
@@ -14,19 +11,6 @@ def my_compiler(gl: paddlefx.GraphLayer, example_inputs: List[paddle.Tensor] = N
     print("my_compiler() called with FX graph:")
     gl.graph.print_tabular()
     return gl.forward
-
-
-def simple_callback(frame: types.FrameType, supported_ops: List[str] = []):
-    # TODO: add a method for frame skiping
-    if frame.f_code.co_name not in supported_ops:
-        return None
-
-    print(frame)
-    print(dis.disassemble(frame.f_code))
-
-    f_code = frame.f_code
-    g = paddlefx.GuardedCode(f_code)
-    return g
 
 
 def add(a, b):
@@ -43,12 +27,5 @@ def func(a=1, b=3):
     return d
 
 
-# paddlefx.optimize
 res = func(1, 3)
-print(res)
-
-# simple_callback
-callback = partial(simple_callback, supported_ops=['add'])
-with paddlefx.DynamoContext(callback):
-    res = add(1, 3)
 print(res)
